requirements.txt
aiogram==3.15.0
google-generativeai==0.8.5
aiosqlite==0.20.0
python-dotenv==1.1.0
